bringup: split SRAM/PSRAM into bringup_sram.c; Step 11 J-Link debug PASS
- Extract sram_test() and psram_test() from bringup_flash.c into new bringup_sram.c
- Add SRAM usage reporting via linker symbols (__data_start__, __bss_end__, __StackTop)
- Reduce rec_buf from 3s to 1s (REC_SECONDS 3→1): static SRAM 313KB→122KB (60%→23%)
- Step 11 complete: register read, memory read, breakpoint, single-step, variable watch all PASS
- Add J-Link Commander script (jlink_step11a.jlink) and GDB scripts (step11b/c.gdb)
- Add send_sram.py serial trigger helper for automated GDB testing
- CLAUDE.md: add Language rule and J-Link device name guidance (RP2350_M33_0/1)
